fix: set_output_mode does not work properly

+++ include/ura/view/output.hpp
@@ -30,7 +30,7 @@ public:
   bool set_mode(sol::table& mode);
   bool set_preferred_mode();
   bool try_set_custom_mode();
-  wlr_output_mode* find_nearest_mode(int width, int height, float refresh);
+  wlr_output_mode* find_nearest_mode(int width, int height, int refresh);
 
   /* Surfaces */
   Vec<UraLayerShell*> bottom_surfaces;
